refactor: organize the server middlewares
