+++ server-middleware/api/donate/invoice.js
@@ -2,7 +2,7 @@ const CryptoJS = require('crypto-js')
 const rawurlencode = require('locutus/php/url/rawurlencode')
 const { post: axiosPost } = require('axios')
 const truncate = require('html-truncate')
-const { EZPAY } = require('../../configs/config')
+const { EZPAY } = require('../../../configs/config')
 
 const CARRIER_TYPE = { PHONE: '0', NATURAL: '1', EXPAY: '2' }
