+++ server-middleware/api/google-sheets/index.js
@@ -8,8 +8,8 @@ const bodyParser = require('koa-bodyparser')
 const Router = require('koa-router')
 const { google } = require('googleapis')
 
-const { reportApiErrorFromKoa } = require('../helpers/index.js')
-const auth = require('./middleware/auth.js')
+const { reportApiErrorFromKoa } = require('../../../helpers/index.js')
+const auth = require('./auth.js')
 
 const app = new Koa()
 const router = new Router()
