+++ server-middleware/api/public.js
@@ -2,9 +2,9 @@ const Koa = require('koa')
 const Router = require('koa-router')
 const { create: axiosCreate } = require('axios')
 
-const { CMS_ENDPOINT_DEPRECATED } = require('../configs/config.js')
-const { getErrorName } = require('../helpers/index.js')
-const { handleKoaCors: handleCors } = require('./middleware/cors.js')
+const { CMS_ENDPOINT_DEPRECATED } = require('../../configs/config.js')
+const { getErrorName } = require('../../helpers/index.js')
+const { handleKoaCors: handleCors } = require('./helpers/cors.js')
 
 const app = new Koa()
 const router = new Router()
