+++ server-middleware/api/subscriptions.js
@@ -14,8 +14,8 @@ const isEmail = require('validator/lib/isEmail')
 
 const { decamelizeKeys } = require('humps')
 const { default: isMobilePhone } = require('validator/lib/isMobilePhone')
-const { CMS_ENDPOINT_DEPRECATED } = require('../configs/config')
-const { handleExpressCors: handleCors } = require('./middleware/cors.js')
+const { CMS_ENDPOINT_DEPRECATED } = require('../../configs/config.js')
+const { handleExpressCors: handleCors } = require('./helpers/cors.js')
 
 const apiHost = CMS_ENDPOINT_DEPRECATED
